ticket:33 Moved UserResource to users app
