+++ p2psafety_django/events/tests/helpers/factories.py
@@ -1,24 +1,8 @@
-from factory import SubFactory, Sequence
+from factory import SubFactory
 from factory.django import DjangoModelFactory, FileField
 
-from django.contrib.auth.models import User
-
 from events.models import Event, EventUpdate
-
-
-class UserFactory(DjangoModelFactory):
-    FACTORY_FOR = User
-    FACTORY_DJANGO_GET_OR_CREATE = ('username',)
-
-    username = Sequence(lambda n: 'user%d' % n)
-    password = Sequence(lambda n: 'user%d' % n)
-
-    @classmethod
-    def _create(cls, target_class, *args, **kwargs):
-        manager = cls._get_manager(target_class)
-        user = manager.create_user(*args, **kwargs)
-        user.real_password = kwargs.get('password')
-        return user
+from users.tests.helpers import UserFactory
 
 
 class EventFactory(DjangoModelFactory):
